Changes in control and file services

+++ UnaCloudFileManager/src/java/uniandes/unacloud/file/net/task/FileReceiverTask.java
@@ -35,7 +35,7 @@ public class FileReceiverTask implements Runnable {
 	private Socket s;	
 	
 	public FileReceiverTask(Socket s) {
-		System.out.println("Attending "+s.getRemoteSocketAddress());
+		System.out.println("Attending " + s.getRemoteSocketAddress());
 		this.s = s;
 	}
 	
@@ -43,35 +43,48 @@ public class FileReceiverTask implements Runnable {
 	public void run() {
 		String mainFile = null;
 		String newMainFile = null;
-		String message;
-		try (Socket ss = s; DataInputStream is = new DataInputStream(s.getInputStream()); Connection con = FileManager.getInstance().getDBConnection();) {
+		String message = null;
+		try (Socket ss = s; DataInputStream is = new DataInputStream(s.getInputStream());) {
 			
 			Long execution = is.readLong();
 			String token= is.readUTF();
-			System.out.println("\tRequest " +execution+" - "+ token);
-			ImageFileEntity image = ImageFileManager.getImageWithFile(token, con);
+			System.out.println("\tRequest " + execution + " - " + token);
+			
+			ImageFileEntity image = null;
+			UserEntity user = null;
+			List<PlatformEntity> platforms = null;
+			try (Connection con = FileManager.getInstance().getDBConnection();) {
+				image = ImageFileManager.getImageWithFile(token, con);
+				if (image != null) {
+					user = UserManager.getUser(image.getOwner().getId(), con);
+					platforms = PlatformManager.getAll(con);
+				}
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			
 			System.out.println("\tImage requested " + image);	
 			
 			if (image != null) {
-				
-				UserEntity user = UserManager.getUser(image.getOwner().getId(), con);
+				boolean received = false;
+				Long sizeImage = 0l;
 				mainFile = image.getRepository().getRoot() + image.getName() + "_" + user.getUsername() + File.separator;//getMainFile().substring(0,image.getMainFile().lastIndexOf(java.io.File.separatorChar)+1);
 				System.out.println("save in path: " + mainFile);
-				TreeMap<File, String>filesTemp = new TreeMap<File, String>();
+				TreeMap<File, String> filesTemp = new TreeMap<File, String>();
 				
 				try (ZipInputStream zis = new ZipInputStream(is)) {
 					
 					System.out.println("\tZip open");
 					final byte[] buffer = new byte[1024 * 100];
 					// for(ZipEntry entry;(entry=zis.getNextEntry())!=null;){					
-					List<PlatformEntity>platforms = PlatformManager.getAll(con);				
+									
 					
 					for (ZipEntry entry; (entry = zis.getNextEntry()) != null;) {
 						boolean goodExtension = false;
 						String mainExtension = null;
 						System.out.println("\t\tFile: " + entry.getName());
 						for (PlatformEntity hyperv : platforms)
-							if (hyperv.validatesExtension(entry.getName())){		
+							if (hyperv.validatesExtension(entry.getName())) {		
 								goodExtension = true;
 								mainExtension = hyperv.getExtension();
 								break;
@@ -79,12 +92,10 @@ public class FileReceiverTask implements Runnable {
 						if (goodExtension) {
 							File tempFile = File.createTempFile(entry.getName(), null);
 							try (FileOutputStream fos = new FileOutputStream(tempFile)) {
-								for (int n; (n = zis.read(buffer)) != -1;) {
+								for (int n; (n = zis.read(buffer)) != -1;)
 									fos.write(buffer, 0, n);
-								}
-								if (entry.getName().endsWith(mainExtension)){
-									newMainFile = mainFile + entry.getName();								
-								}								
+								if (entry.getName().endsWith(mainExtension))
+									newMainFile = mainFile + entry.getName();
 							}	
 							System.out.println("Save temp "+tempFile);
 							filesTemp.put(tempFile,entry.getName());
@@ -93,7 +104,7 @@ public class FileReceiverTask implements Runnable {
 						zis.closeEntry();
 					}
 					System.out.println("There are "+filesTemp.size() + ", Delete old files");
-					Long sizeImage = 0l;
+					
 					
 					if (filesTemp.size() > 0) {			
 						
@@ -110,36 +121,37 @@ public class FileReceiverTask implements Runnable {
 							//newFile.createNewFile();
 							System.out.println("save: "+newFile);
 							try (FileInputStream streamTemp = new FileInputStream(temp); FileOutputStream ouFile = new FileOutputStream(newFile)) {
-								for (int n; (n = streamTemp.read(buffer)) != -1;) {
-									ouFile.write(buffer, 0, n);
-								}															
+								for (int n; (n = streamTemp.read(buffer)) != -1;)
+									ouFile.write(buffer, 0, n);																						
 							}	
 							
 							sizeImage += temp.length();
 							temp.delete();
 						}
 					}				
-					
-					System.out.println("reception finished: " + newMainFile);
-					try {
-						ImageFileManager.setImageFile(new ImageFileEntity(image.getId(), ImageEnum.AVAILABLE, null, null, null, null, sizeImage, newMainFile, null, null), false, con, false);
-						message = "Image has been saved in server";
-						System.out.println("Status changed, process closed");
-					} catch (Exception e) {
-						e.printStackTrace();
-						 message = e.getMessage();
-					}
-					
+					received = true;
+					System.out.println("reception finished: " + newMainFile);	
+					message = "Image has been saved in server";
 				} catch (Exception e) {		
 				    e.printStackTrace();
-				    ImageFileManager.setImageFile(new ImageFileEntity(image.getId(), ImageEnum.UNAVAILABLE, null, null, null, null, null, null, null, null), false, con, false);
 				    message = e.getMessage();
-				    for (File tmpFile : filesTemp.keySet()) {
-						tmpFile.delete();
-					}
 				}	
-				DeploymentManager.setExecution(new ExecutionEntity(execution, 0, 0, null, new Date(), null, ExecutionStateEnum.FINISHED, null, message), con);
-				DeploymentManager.breakFreeInterfaces(execution, con, IPEnum.AVAILABLE);
+				
+				try (Connection con = FileManager.getInstance().getDBConnection();) {
+					if (received) {
+						ImageFileManager.setImageFile(new ImageFileEntity(image.getId(), ImageEnum.AVAILABLE, null, null, null, null, sizeImage, newMainFile, null, null), false, con, false);
+						System.out.println("Status changed, process closed");
+					} else {
+						 ImageFileManager.setImageFile(new ImageFileEntity(image.getId(), ImageEnum.UNAVAILABLE, null, null, null, null, null, null, null, null), false, con, false);
+						 System.out.println("Error in process, all files should be deleted");
+						 for (File tmpFile : filesTemp.keySet())
+							tmpFile.delete();	
+					}
+					DeploymentManager.setExecution(new ExecutionEntity(execution, 0, 0, null, new Date(), null, ExecutionStateEnum.FINISHED, null, message), con);
+					DeploymentManager.breakFreeInterfaces(execution, con, IPEnum.AVAILABLE);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
